Menu Customizer: Build-out the ajax potion of the add-menu-item process.
Items are sent one-at-a-time, as they're now added one-at-a-time. A new function handles the initial process, then it is fed into and back out of the existing system `wp_save_menu_items()`.

`WP_Menu_Item_Customize_Control::render_content()` has been extracted to a function that is also used to prepare the html for the new item's control, which is returned to the page and rendered in place of the placeholder.

Still to do: register the new item's setting and control in JS, fix ajax handling of custom items, and tweak various pieces of UI-oriented JS to account for dynamically added items.

git-svn-id: http://plugins.svn.wordpress.org/menu-customizer/trunk@935321 b8457f37-d9ea-0310-8a92-e5e31aec5664