Menu Customizer: use the JS/Underscore template for menu item controls that are dynamically added, eliminating the need for `menu_customizer_render_item_control()`. Still working out some kinks with the new JS API, including handling of the active state when adding the controls dynamically.
Note that it is still necessary to use Ajax in this process because menu items must be created in the DB to be previewed with the current system. This allows the code that prepares existing menu items to be rendered to also be used for new menu items, by creating an instance of `WP_Customize_Menu_Item_Control` and calling `json()`.

See #2278.

git-svn-id: http://plugins.svn.wordpress.org/menu-customizer/trunk@1043169 b8457f37-d9ea-0310-8a92-e5e31aec5664